Create README.md I'm interested in learning the best code s and be apart of a good program. I'm currently learning at my own level of understanding without help , if I can help some out on a project reach out to me @ sammiehoss30@gmail.com.. thanks
Basically I'm interested in learning more than basic stuff, learning the inside of the process of building a new setup and having success with my project. 
If I could tag along with a current project or program and get a feel for it , an idea of the direction I want to go will come better into focus and I can get my own program off the ground.